requirements.txt
openai>=1.30.0
requests>=2.31.0
sounddevice>=0.4.6
webrtcvad>=2.0.10
numpy>=1.26.0
pygame>=2.5.2
python-dotenv>=1.0.1
